requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-dotenv==1.0.0
nltk==3.8.1
schedule==1.2.1
selenium==4.16.0
webdriver-manager==4.0.1 